Add files via upload
Code of the first assignment.
Simple operations in homogeneous coordinates
Finding the points of infinity of a square and the the line of infinity. Then finding the points of intersection of the square diagonals with the line of infinity.
Assuming pinhole camera we are using the collinearity condition to simulate the photograph taken of the square with a camera 4912x3264 pixel. After that we calculate the vanishing points of the square in the simulated photograph.